Update patch set 1
Patch Set 1:

> I expected to see new usages of go:nowritebarrierrec. Can we add it
 > to gentraceback? Why?

We could, but the constraint isn't on gentraceback, it's on some of the callers of gentraceback. go:nowritebarrierrec is meant to mean "there must not be write barriers", not "there don't happen to be write barriers". The go:nowritebarrierrec where this matters is introduced in CL 17006 after a few more fixes.

Patch-set: 1