cmd/dist: add -buildmode flag for testing PIE
A builder can run all.bash with -buildmode=pie and test the standard
library with PIE binaries. This is interesting because 386 assembly
behaves differently in position independent code (the CX register
can be borrowed, see the discussion on http://golang.org/cl/16385).

There are three tests that don't pass on linux with this

	runtime/trace: TestTraceSymbolize
	cmd/addr2line: TestAddr2Line
	cmd/nm: TestNM

If this approach to a PIE builder is acceptable, I'll take a look at
them in a followup CL.

Updates #13136

Change-Id: Ie70bda9afb885a314cf9882cf9f3934c0d78c77d