Update patch set 2
Patch Set 2: Code-Review-2

Thanks but no.

While it's true that some regexp packages do accept [-] or [x-] as equivalent to [\-] or [x\-], it's not universal. Go is not adhering to some external spec here. We defined a specific grammar (see the doc comment on filepath.Match) and that grammar does not admit [-] nor [x-].

Patch-set: 2
Label: Code-Review=-2