Update patch set 1
Patch Set 1: Code-Review-1

I've tested this code with the original test suite which caused me to find the issue (the python math test suite).

It fixes most of the issues, but not all

math.Gamma(-1000.5) should be -0.0 but is returned as NaN

math.Gamma(-1000000000.3) should be -0.0 but is returned as NaN

Sorry!

Nick

Patch-set: 1
Reviewer: Gerrit User 5175 <5175@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1