net/http: fix race when request body is reused before being read to EOF
When a server replies with an HTTP response before the HTTP request
body is fully written, the net/http package now wraps the response
body with a version which when done (at close or the final read) will
wait (for a bounded amount of time) for the the request write to
terminate.

Assuming the writing is still making progress (doing normal Read+Write
calls from the request body to the underlying network socket), the
writes will start to fail as the persistent connection is closed in
these cases. If it's not, we give the reader a bit of time to
terminate and then stop using it.

This also fixes the flakiness and speed of another non-short test I
noticed while testing.

Fixes #12796

Change-Id: I82b50b405f5e06e39f44e0284cc83c9aa46fb254