Update patch set 4
Patch Set 4:

Yes, there is a need to set the name of the shared library, but I think, that it should not be done as a part of this cl, which just fixes the issue.
Are you suggesting removing of this limit? If so, then it can be simply set to 255 symbols, because longer names are not possible, and explicit lib names (when implemented) can be used in extraordinary cases.

Patch-set: 4