Update patch set 9
Patch Set 9: Code-Review-2

The right reviewer is Mike Samuel, but he's hard to get a hold of. 

There are serious security implications to thinking <script type="text/javascript"> is the only <script> tag that contains JavaScript. I am very skeptical that modern browsers are that disciplined, although the discussion on #6701 suggests that maybe reacting to type= is OK.

Also I've never seen <script type="text/template">. What is that? Is it supported by the standard library?

Patch-set: 9
Label: Code-Review=-2