Update patch set 3
Patch Set 3: -Run-TryBot

> Thank you for doing this, but I propose you review my CL 16854
 > first.  Only because it adds test that verifies net interface
 > addresses and masks and compares them with output of windows netsh
 > command. This should give us some assurance that we're moving in
 > the right direction. My test only includes ipv4 addresses and does
 > not work on winxp, but is better than nothing.
 > 
 > What do you think?

i think it's better to land this cl first because this cl doesn't contain complicated bug fixes.

just skimmed cl 16854, i agree that having windows-specific tests sounds reasonable, but the cl tries to find out on-link prefixes by inference.

as described in issue 11837 and 13360 a bit, i think that the current commonPrefixLen is not suitable for generating concrete information on ip routing because it just picks perhaps better one for dialing w/ multiple destinations and it's enough for that purpose; inaccuracy doesn't matter because dial tries the next one. on the other hand, interface address prefixes are equivalent to connected routes and it should be concrete and accurate for helping applications such as routing and singnaling protocols.

i'm not sure whether we should go w/ cl 16854 during go1.6 cycle.

Patch-set: 3
Label: Run-TryBot=0