Update patch set 1
Patch Set 1:

Do we want to go down this road and document every attack Go does not defend against? From the tweet (https://twitter.com/agl__/status/669182140244824064) it seems like agl's position is that the cure is worse than the disease in this case. 

I might be persuaded if we knew that other common servers did turn on Lucky13 hardening (whatever that means) by default. Is that the case?

Patch-set: 1