spec: proposal: require 16 bit minimum exponent in constants rather than 32
THIS IS A LANGUAGE CHANGE.

A 16bit binary exponent permits a constant range covering e-32768 to 1e+32767
which is more than enough for practical and hypothetical constant arithmetic.

Furthermore, until recently cmd/compile could not handle very large exponents
correctly anyway; i.e., the chance that any real programs (but for tests that
explore corner cases) are affected are close to zero.

Finally, restricting the minimum supported range significantly reduces the
implementation complexity in an area that hardly matters in reality for new
or alternative spec-compliant implementations that don't or cannot rely on
pre-existing arbitratry precision arithmetic packages that support a 32bit
exponent range.

Change-Id: I970f919c57fc82c0175844364cf48ea335f17d39