cmd/cgo: ISO standard C preamble for complex types
The existing complex typedefs in the CGO preamble use the __complex keyword,
which is a GNU extension to C, not official ISO standard. Most compilers can
work with this, but some (e.g. SWIG) choke on it as syntax error.

This is a similar change to the one merged into libgo about a year ago by
Ian: https://gcc.gnu.org/ml/gcc-patches/2014-05/msg02510.html

Change-Id: I3e7a38cdbee58757ab030b63c7089b02f797649a