Update patch set 2
Patch Set 2:

The commit message claims 16-bit binary exponent corresponds to 1e-32768 to 1e+32767 but it does not: those are decimal exponents, not binary. (I've checked in code making this mistake in the past.)

http://www.wolframalpha.com/input/?i=2%5E32767 and
http://www.wolframalpha.com/input/?i=2%5E-32768

says that the range is roughly 7e-9865 to 7e9863.

I am in favor of this change. This eliminates a corner of the language that historically has been implemented very badly, if at all. Although the current implementations are slowly coming to grips with 32-bit exponents, the path has been long and rocky and the benefit seemingly non-existent, suggesting that we should relieve both current and future implementations of these woes. 

In Go 1.4 the compiler rejected exponents larger than 10000 (due to knowing the code didn't work for larger exponents) without any complaint from users. In earlier versions of Go, large exponents were silently mishandled, again without any complaint from users.

Patch-set: 2
Label: Code-Review=0