Update patch set 2
Patch Set 2: Code-Review-1

While I sympathize with the sentiment and am sure to be overruled, I don't like this change because of the precedent it sets. This change is triggered by an inconvenience created by a new implementation of constants in the compiler, and rather than grapple with the requirements of the specification, the proposed remedy is to change the language. There's really no other way to spin this: we propose to change the language to simplify the implementation of the compiler. That may be a valid thing to do but it gives me a bad taste.

Patch-set: 2
Label: Code-Review=-1