Update patch set 4
Patch Set 4:

I side with Ian here. This is a valuable bug fix that repairs the code to work as documented and makes it easier in general to use and more regular.

No need to roll it back. You just need to adjust your habits.

(Besides, even if this weren't a bug fix, there is no compatibility requirement for tools.)

Patch-set: 4
Label: Code-Review=0