Update patch set 4
Patch Set 4:

1. It no longer behaves as documented. "[flags for test binary]" was meant to include things like -short. "go test math -short" no longer works.

2. The docs do explain the old behavior:

The test binary also accepts flags that control execution of the test; these
flags are also accessible by 'go test'. See 'go help testflag' for details.

If the test binary needs any other flags, they should be presented after the
package names. The go tool treats as a flag the first argument that begins with
a minus sign that it does not recognize itself; that argument and all subsequent
arguments are passed as arguments to the test binary.

3. It breaks people's scripts. It even broke cmd/dist, which someone then fixed in cmd/dist instead of rolling back this CL.

4. It breaks people's fingers. I've been confused by this since I returned from paternity leave, but due to a combination of not spending much time coding and not trusting my brain due to paternity-induced sleep loss, I had not progressed from being mystified as to what's going on to realizing it's an actual regression. I thought I was just confused. This is sure to bite others.

Patch-set: 4