Update patch set 4
Patch Set 4:

Even assuming the new behavior is superior to the old and fixes a problem (neither of which I'm convinced of - this comes up in exactly one package, glog), this contradicts the documented behavior of the last 6 releases.

The docs say (even today):

---
test [-c] [-i] [build and test flags] [packages] [flags for test binary]
---

It's certainly the case that -short is a flag for the test binary: the only effect it has is to be passed to the test binary, albeit as -test.short. It's handled in either position to avoid confusion. (More below about how this came to be.)

I admit this is not 100% clear, but it's hard to be 100% clear in a single punched card-sized line.  That's why the rest of the docs elaborate.

The docs go on:

---
The test binary also accepts flags that control execution of the test; these
flags are also accessible by 'go test'. See 'go help testflag' for details.

If the test binary needs any other flags, they should be presented after the
package names. The go tool treats as a flag the first argument that begins with
a minus sign that it does not recognize itself; that argument and all subsequent
arguments are passed as arguments to the test binary.
---

and then 'go help testflag' mentions flags like -short and says:

---
The following flags are recognized by the 'go test' command and
control the execution of any test:

...

The test binary, called pkg.test where pkg is the name of the
directory containing the package sources, can be invoked directly
after building it with 'go test -c'. When invoking the test binary
directly, each of the standard flag names must be prefixed with 'test.',
as in -test.run=TestMyFunc or -test.v.

When running 'go test', flags not listed above are passed through
unaltered. For instance, the command

	go test -x -v -cpuprofile=prof.out -dir=testdata -update

will compile the test binary and then run it as

	pkg.test -test.v -test.cpuprofile=prof.out -dir=testdata -update

The test flags that generate profiles (other than for coverage) also
leave the test binary in pkg.test for use when analyzing the profiles.

Flags not recognized by 'go test' must be placed after any specified packages.
---

It does not say that flags recognized by 'go test' must be placed BEFORE the packages, and our own scripts have not done this.

Literally the very first checkin of 'go test' showed in its commit message the syntax being removed in this CL: "go test all -short". https://codereview.appspot.com/5495055

It's very confusing to explain why (after the current CL) one can write 

go test -test.short all
go test all -test.short
go test -short all
but not 
go test all -short

Instead of explaining that, we made the fourth case work too. This was intentional, it was documented, and people now rely on it, and have for years.

As golang.org/issue/7221 comment 2 says, if one must pass a literal -v to a test binary, it is possible to do so by giving the binary some no-op flag (say, -nop), which will be unrecognized by go test and terminate the flag conversion processing. Then "go test -nop -v" will get passed to the binary as "-nop -v". This solution works even when testing in the current directory without naming the package, instead of overloading the presence of the package name. If we need to do something explicit, we can add some kind of -stop-reading-flags flag.

I'm concerned that we're breaking years of established usage for exactly one test case, package glog. I understand that you didn't use the go command that way, but many people did and still do.

Patch-set: 4