Update patch set 4
Patch Set 4:

> You are the first to complain about this change, so I really don't
 > believe it's a problem. Instead, I think it fixes a real problem
 > that has tripped up a number of people and has been reported
 > several times.

I think I'm the first to complain because everyone working at tip just accepted the breakage it caused. But that doesn't mean its right, and it doesn't mean it won't cause many more problems when it hits a real release. I mentioned having found this CL to Austin this morning and he indicated that he'd hit this problem too, but you seemed so sure it was a correct change that it didn't seem worth objecting.

Patch-set: 4