Update patch set 2
Patch Set 2:

In general there is inconsistency on Unix programs about where the output of -help should be printed. Often it is printed to standard output, in contrast to errors. The theory is that when you say -help you are expecting the output, so it is not an error. I'm happy to leave this for now, but if we get reports that it broke something, we may have to roll it back.

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0