Update patch set 4
Patch Set 4:

The workaround for the original problem is distasteful.

It's clear you want to roll this back over my objections. That's fine. I think that it's a mistake and that the old behavior was confusing for me and many others. It has come up in other cases and the old behavior is just too special. Nothing else I've used works like that.

I'm not going to press my point any further but please reopen the bug about -v or make the documentation a lot clearer.

Patch-set: 4