Update patch set 2
Patch Set 2:

This seems OK, but can we make the buildmode not mentioned unless asked for? That would mean making the flag default to "", not passing it on the command lines when "", and not mentioning it in the header when "".

As much as possible I would like to continue to let people ignore all the complexity creeping in here.

Otherwise I am happy with this and would like to see it go in if you are still interested.

Patch-set: 2