Update patch set 1
Patch Set 1: Code-Review+2

I was going to do this too but was a bit worried that it hasn't been published yet, despite being approved. But this is fine. It's in use.

We've already api/go1.6.txt to update api/go1.6.txt in this CL too and kick off the trybots to make sure you got it.

Patch-set: 1
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2