Use the getentropy syscall on OpenBSD
Go already supports Linux's getrandom, which is a slightly modified
version of getentropy.

All supported versions of OpenBSD include getentropy so, unlike with
Linux and getrandom, we don't need to test for its presence.

Solaris added tweaked versions of getentropy and getrandom in July 2015,
so this may be relevant to Go's Solaris support as well.

Resolves #13785

Change-Id: Ib536b96675f257cd8c5de1e3a36165e15c9abac9