cmd/compile: better modeling of escape from closures
Even when maps are captured "by value" their internal
data structures are still shared.  If the capturing closure
ends up heap allocated, the heap "value" within ends up
pointing to stack memory.  Stir in some recursion to induce
stack reallocation (and movement) and disaster ensues.

This turns out to generalize.  Something as simple as
  var x int
  m := &x
  ...
  fn = func() {
	if j < 100 {
		j++
		fn()
	} else {
		*m = *m + 1
	}
  }
can also trigger this failure.

The main approach to the fix is to make heap-escape
taint everything downstream in the "flood" phase.
This has the flavor of "how could this not already be
the case?" but in fact it was not -- though the existing
escape analysis tests were unperturbed by this modification.

The new test checks the following:
1) in the map case, that if fn escapes, so does the map.
2) in the map case, if fn does not escape, neither does the map.
3) in the &x case, that if fn escapes, so does &x.
4) in the &x case, if fn does not escape, neither does &x.

Fixes #13799.

Change-Id: Ie280bef2bb86ec869c7c206789d0b68f080c3fdb