cmd/compile: tweaked escape of closure-captured maps
Even when maps are captured "by value" their internal
data structures are still shared.  If the capturing closure
ends up heap allocated, the heap "value" within ends up
pointing to stack memory.  Stir in some recursion to induce
stack reallocation (and movement) and disaster ensues.

The "fix" is two-part: (1) if a captured variable is a map,
treated it as addressed even if it is by-value; (2) in the
"flood" phase, make it possible to enforce the rule that
by default a reference from heap-allocated anything is
leak-inducing.  Both changes are conservative, both pass
tests, and relaxing either one leads to failure.  If the
test is modified to avoid heap allocation of the closure,
the the map is also not heap allocated, so it appears to not
be overconservative.

It seems possible that a similar escape-analysis gotcha
could be constructed with slices; if that is the case,
then they receive the special treatment in closure-capture
code.

Fixes #13799.

Change-Id: Ie280bef2bb86ec869c7c206789d0b68f080c3fdb