cmd/compile: better modeling of escape across loop levels
Includes a bit mask we can fiddle with to test escape
enhancement.

Even when maps are captured "by value" their internal
data structures are still shared.  If the capturing closure
ends up heap allocated, the heap "value" within ends up
pointing to stack memory.  Stir in some recursion to induce
stack reallocation (and movement) and disaster ensues.

This turns out to generalize.  Something as simple as
  var x int
  m := &x
  ...
  fn = func() {
	if j < 100 {
		j++
		fn()
	} else {
		*m = *m + 1
	}
  }
can also trigger this failure.

And it also generalizes to not-closures; the problem involves
accounting for escapes induced by leaks out of loop levels.
These are "capacity" escapes, not "publishing" escapes

The fix is that when a heap allocation is newly discovered,
use its looplevel as the minimum loop level for downstream
escape flooding.

The new test checks the following:
1) in the map case, that if fn escapes, so does the map.
2) in the map case, if fn does not escape, neither does the map.
3) in the &x case, that if fn escapes, so does &x.
4) in the &x case, if fn does not escape, neither does &x.

Fixes #13799.

Change-Id: Ie280bef2bb86ec869c7c206789d0b68f080c3fdb