Update patch set 2
Patch Set 2:

This is not the correct fix. The new semantics of go/constant doesn't convert between Kind types automatically anymore; one has to call ToInt/ToFloat/ToComplex as needed. Your change works accidentally because the numbers are small and thus become rationals. More generally, a large number could become a Float and then your change wouldn't work.

Please abandon. This is not a bug in go/constant but - if anywhere - in the code using go/constant.

Patch-set: 2