Update patch set 2
Patch Set 2:

> According to http://build.golang.org, this CL has started causing
 > the runtime/pprof test to fail on Solaris and Linux/ARM.

On linux/arm, I'm sure this just increased the probability of https://golang.org/issue/13405. It wouldn't surprise me too much if Solaris has a similar problem.

This is a real issue with profiling on these platforms, but maybe we should disable this test for now to clear up the build.

Patch-set: 2