Update patch set 3
Patch Set 3: Code-Review-2

R=go1.7

Thanks very much. As a developer, I'm happy to see these changes but now is not the right time.

Code cleanup has no benefit to end users; it can only break things, and there is not much time left to find bugs introduced. We can land this CL early in the Go 1.7 cycle but not at this point in the Go 1.6 cycle.

Patch-set: 3